[ski ci] refactor
