+++ benchmarks/run_benchmark.py
@@ -1,4 +1,3 @@
-import argparse
 import sys
 import logging
 import subprocess
@@ -10,35 +9,20 @@ from torchvision.prototype.datasets import load
 import torch.nn.functional as F
 from torchvision import transforms
 import time
-from statistics import mean
 import torch.optim as optim
 import torch.profiler
+
+# Relative imports
+from args import arg_parser
+from utils import init_fn
+from datasets import prepare_gtsrb
+from trainers import train
+from report import create_report
 
 logging.basicConfig(filename='example.log', level=logging.DEBUG)
 
 
-## Arg parsing
-parser = argparse.ArgumentParser()
-parser.add_argument("--dataset", type=str, default="gtsrb", help="The name of the dataset")
-parser.add_argument("--model_name", type=str, default="resnext50_32x4d", help="The name of the model")
-parser.add_argument("--batch_size", type=int, default=1, help="")
-parser.add_argument("--device", type=str, default="cuda:0", help="Options are are cpu or cuda:0")
-parser.add_argument("--num_epochs", type=int, default=1)
-parser.add_argument("--report_location", type=str, default="./report.md", help="The location where the generated report will be stored")
-parser.add_argument("--num_workers", type=int, default=1, help="Number of dataloader workers")
-parser.add_argument("--shuffle", action="store_true")
-parser.add_argument("--dataloaderv", type=int, default=1)
-
-args = parser.parse_args()
-dataset = args.dataset
-model_name = args.model_name
-batch_size = args.batch_size
-device = args.device
-num_epochs = args.num_epochs
-report_location = args.report_location
-num_workers = args.num_workers
-shuffle = args.shuffle
-dataloaderv = args.dataloaderv
+dataset, model_name, batch_size, device, num_epochs, num_workers, shuffle, dataloaderv = arg_parser()
 
 if dataloaderv == 1:
     from torch.utils.data import DataLoader
@@ -47,12 +31,7 @@ elif dataloaderv == 2:
 else:
     raise(f"dataloaderv{dataloaderv} is not a valid option")
 
-# Util function for multiprocessing
-def init_fn(worker_id):
-    info = torch.utils.data.get_worker_info()
-    num_workers = info.num_workers
-    datapipe = info.dataset
-    torch.utils.data.graph_settings.apply_sharding(datapipe, num_workers, worker_id)
+
 
 # Download model
 model_map = {
@@ -87,30 +66,7 @@ print(f"Dataset length {len(dp)}")
 logging.debug(f"data format before preprocessing is {next(iter(dp))}")
 
 if dataset == "gtsrb":
-    def transform(img):
-        t= transforms.Compose([
-            transforms.ToPILImage(),
-            transforms.Resize(size=(100,100)),
-            # transforms.reshape(64,3,7,7), TODO: Figure out collation
-            transforms.ToTensor()]
-        )
-        return t(img).to(torch.device(device))
-
-    def str_to_list(str):
-        l = []
-        for char in str:
-            l.append(int(char))
-        return l
-
-    # Filter out bounding box and path to image
-    dp = dp.map(lambda sample : {"image" : sample["image"], "label" : sample["label"]})
-
-    # Apply image preprocessing
-    dp = dp.map(lambda sample : transform(sample.decode()), input_col="image")
-    dp = dp.map(lambda sample : torch.tensor(str_to_list(sample.to_categories())).to(torch.device(device)), input_col="label")
-
-    # Batch
-    dp = dp.batch(batch_size)
+    dp = prepare_gtsrb(batch_size, device, dp)
     
 # Datapipe format after preprocessing
 logging.debug(f"data format after preprocessing is \n {next(iter(dp))}\n")
@@ -123,13 +79,16 @@ if num_workers == 1:
 else:
     dl = DataLoader(dataset=dp, batch_size=batch_size, shuffle=True, num_workers=num_workers, worker_init_fn=init_fn, multiprocessing_context="spawn")
 
+criterion = torch.nn.CrossEntropyLoss()
+optimizer = torch.optim.SGD(model.parameters(), lr=0.001, momentum=0.9)
+
+
 
 total_start = time.time()
 per_epoch_durations = []
 batch_durations = []
 
-criterion = torch.nn.CrossEntropyLoss()
-optimizer = torch.optim.SGD(model.parameters(), lr=0.001, momentum=0.9)
+
 
 with torch.profiler.profile(
     activities=[torch.profiler.ProfilerActivity.CPU, torch.profiler.ProfilerActivity.CUDA],
@@ -142,40 +101,13 @@ with torch.profiler.profile(
     with_modules=True
 ) as p:
 
-    for epoch in range(num_epochs):
-        epoch_start = time.time()
-        running_loss = 0
-        for i, elem in enumerate(dl):
-            batch_start = time.time()
-
-            labels = torch.argmax(elem[0]["label"], dim=1)      
-            optimizer.zero_grad()
-            outputs = model(elem[0]["image"])
-            loss = criterion(outputs,labels)
-            loss.backward()
-            optimizer.step()
-
-            running_loss += loss.item()
-            # if i % 200 == 1999:    # print every 2000 mini-batches
-            print(f'[{epoch + 1}, {i + 1:5d}] loss: {running_loss / 2000:.10f}')
-            running_loss = 0.0
-
-            batch_end = time.time()
-            batch_duration = batch_end - batch_start 
-            batch_durations.append(batch_duration)
-            p.step()
-
-        epoch_end = time.time()
-        epoch_duration = epoch_end - epoch_start
-        per_epoch_durations.append(epoch_duration)
+    train(num_epochs, model, dl, per_epoch_durations, batch_durations, criterion, optimizer, p)
 
     total_end = time.time()
     total_duration = total_end - total_start
 
 # TODO: Make this output some human readable markdown file
-def create_report(per_epoch_durations, batch_durations, total_duration):
-    print(f"Total duration is {total_duration}")
-    print(f"Per epoch duration {mean(per_epoch_durations)}")
-    print(f"Per batch duration {mean(batch_durations)}")
 
 create_report(per_epoch_durations, batch_durations, total_duration)
+
+
